cambiado logica de bd a clase dao, falta adaptar para recibir clase
